verifying if user is authenticated

+++ src/actions/delete-post-action.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { verifyLoginSession } from "@/lib/login/manage-login";
 import { postRepository } from "@/repositories/post";
 import { revalidateTag } from "next/cache";
 
@@ -7,6 +8,14 @@ export const deletePostAction = async (id: string) => {
   if (!id || typeof id !== "string") return { error: "Dados inválidos" };
 
   let post;
+
+  const isAuthenticated = await verifyLoginSession();
+
+  if (!isAuthenticated) {
+    return {
+      errors: "Você precisa estar logado para fazer isso",
+    };
+  }
   try {
     post = await postRepository.delete(id);
   } catch (e) {
